simplify settings.js
