+++ src/lib/editor.js
@@ -1,0 +1,9 @@
+import { editorCopy } from './editorCopy';
+import { editorCreate } from './editorCreate';
+import { editorSave } from './editorSave';
+
+export function editor() {
+	const editor = editorCreate();
+	editorSave(editor);
+	editorCopy(editor);
+}
