+++ src/marker.js
@@ -2,3 +2,4 @@ import { markerCreate } from './lib/markerCreate';
 import { settingsUse } from './lib/settingsUse';
 
 settingsUse(markerCreate);
+
