+++ src/settings.js
@@ -1,15 +1,5 @@
-import { editorCopy } from './lib/editorCopy';
-import { editorCreate } from './lib/editorCreate';
-import { editorSave } from './lib/editorSave';
-import { permissionsCheck } from './lib/permissionsCheck';
-import { permissionsRequest } from './lib/permissionsRequest';
-import { permissionsToggle } from './lib/permissionsToggle';
-import { permissionsWatch } from './lib/permissionsWatch';
+import { editor } from './lib/editor';
+import { permissions } from './lib/permissions';
 
-permissionsCheck(permissionsToggle);
-permissionsWatch(permissionsToggle);
-permissionsRequest();
-const editor = editorCreate();
-editorSave(editor);
-editorCopy(editor);
-
+permissions();
+editor();
